+++ src/DateTime/Clocks/OffsetClock.php
@@ -2,6 +2,7 @@
 namespace Praline\DateTime\Clocks;
 
 use Praline\DateTime\ClockInterface;
+use Praline\DateTime\Duration;
 
 // 將一個時鐘的時間加上偏差值，可用於伺服環境的測試
 class OffsetClock implements ClockInterface
@@ -23,6 +24,16 @@ class OffsetClock implements ClockInterface
         } else {
             throw new \TypeError('Not a string or DateInterval');
         }
+    }
+
+    public static function byWeeks(ClockInterface $referenceClock, int $weeks): OffsetClock
+    {
+        return new OffsetClock($referenceClock, Duration::weeks($weeks));
+    }
+
+    public static function byDays(ClockInterface $referenceClock, int $days): OffsetClock
+    {
+        return new OffsetClock($referenceClock, Duration::days($days));
     }
 
     public function now(): \DateTime